+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-# v0.2.0
+# v0.3.0
 
 ## Enhancements
 
-* Additional registry support.
-* Now supports the ability to scan filesystem for signed executables (EXE) and Java applications (JAR)
+* Registry compatibility improvements.
+* Windows support
+* Experimental SBOM/artifact signature discovery